old-tests/submission/W3C/domhtml/table51.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - TABLECOL</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-col-element.html#dom-col-span" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<table summary="Table Summary">
<colgroup align="center" char="$" charoff="15" span="2" valign="middle" width="20">
<col align="center" char="*" charoff="20" span="1" valign="middle" width="20">
</colgroup>
<tbody><tr>
<th>Id</th>
<th>Name</th>
<th>Position</th>
<th>Salary</th>
</tr>
<tr>
<td>EMP0001</td>
<td>Martin</td>
<td>Accountant</td>
<td>56,000</td>
</tr>
</tbody></table>











<script>
var domtest = async_test("Indicates the number of columns in a group or affected by a grouping.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vspan;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("col");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vspan = testNode.span;

      assert_equals(vspan,1);
       
    domtest.done();
}));

</script>
<p>Indicates the number of columns in a group or affected by a grouping.
The value of attribute span of the tablecol element is read and checked against the expected value.</p>
<div id="log"></div>
</body></html>
